reorganise hooks
