+++ packages/block-editor/src/components/inserter/block-patterns-explorer/hooks/index.js
@@ -1,0 +1,3 @@
+export { default as useDebouncedInput } from './use-debounced-input';
+export { default as usePatternsFromDirectory } from './use-patterns-from-directory';
+export { default as usePatternDirectoryCategories } from './use-pattern-directory-categories';
