+++ packages/block-editor/src/components/inserter/block-patterns-explorer/hooks/use-pattern-directory-categories.js
@@ -1,0 +1,17 @@
+/**
+ * WordPress dependencies
+ */
+import { useState, useEffect } from '@wordpress/element';
+import apiFetch from '@wordpress/api-fetch';
+
+export default function usePatternDirectoryCategories() {
+	const [ categories, setCategories ] = useState( [] );
+	useEffect( () => {
+		apiFetch( {
+			path: '/wp/v2/pattern-directory/categories',
+		} ).then( ( fetchedCategories ) => {
+			setCategories( fetchedCategories );
+		} );
+	}, [] );
+	return categories;
+}
