+++ packages/block-editor/src/components/inserter/block-patterns-explorer/hooks/use-patterns-from-directory.js
